Extract RandomChooser class
